validators: Reset subtopic to original state and add missing topic validator,test to appropriate file
