+++ core/domain/prod_validators.py
@@ -48,6 +48,9 @@ from core.domain import story_fetchers
 from core.domain import subtopic_page_domain
 from core.domain import subtopic_page_services
 from core.domain import suggestion_services
+from core.domain import topic_domain
+from core.domain import topic_fetchers
+from core.domain import topic_services
 from core.domain import user_domain
 from core.domain import user_services
 from core.domain import voiceover_services
@@ -2206,51 +2209,6 @@ class ExplorationRecommendationsModelValidator(
         return [cls._validate_item_id_not_in_recommended_exploration_ids]
 
 
-class TopicSimilaritiesModelValidator(base_model_validators.BaseModelValidator):
-    """Class for validating TopicSimilaritiesModel."""
-
-    @classmethod
-    def _get_model_id_regex(cls, unused_item):
-        # Valid id: topics.
-        return '^%s$' % recommendations_models.TOPIC_SIMILARITIES_ID
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return []
-
-    @classmethod
-    def _validate_topic_similarities(cls, item):
-        """Validate the topic similarities to be symmetric and have real
-        values between 0.0 and 1.0.
-
-        Args:
-            item: datastore_services.Model. TopicSimilaritiesModel to validate.
-        """
-
-        topics = list(item.content.keys())
-        data = '%s\n' % (',').join(topics)
-
-        for topic1 in topics:
-            similarity_list = []
-            for topic2 in item.content[topic1]:
-                similarity_list.append(
-                    python_utils.UNICODE(item.content[topic1][topic2]))
-            if len(similarity_list):
-                data = data + '%s\n' % (',').join(similarity_list)
-
-        try:
-            recommendations_services.validate_topic_similarities(data)
-        except Exception as e:
-            cls._add_error(
-                'topic similarity check',
-                'Entity id %s: Topic similarity validation for content: %s '
-                'fails with error: %s' % (item.id, item.content, e))
-
-    @classmethod
-    def _get_custom_validation_functions(cls):
-        return [cls._validate_topic_similarities]
-
-
 class StoryModelValidator(base_model_validators.BaseModelValidator):
     """Class for validating StoryModel."""
 
@@ -2941,6 +2899,124 @@ class CommunityContributionStatsModelValidator(
             cls._validate_question_reviewer_count,
             cls._validate_question_suggestion_count
         ]
+
+
+class SubtopicPageModelValidator(base_model_validators.BaseModelValidator):
+    """Class for validating SubtopicPageModel."""
+
+    @classmethod
+    def _get_model_id_regex(cls, item):
+        return '^%s-\\d*$' % (item.topic_id)
+
+    @classmethod
+    def _get_model_domain_object_instance(cls, item):
+        return subtopic_page_services.get_subtopic_page_from_model(item)
+
+    @classmethod
+    def _get_external_id_relationships(cls, item):
+        snapshot_model_ids = [
+            '%s-%d' % (item.id, version) for version in python_utils.RANGE(
+                1, item.version + 1)]
+        return [
+            base_model_validators.ExternalModelFetcherDetails(
+                'subtopic_page_commit_log_entry_ids',
+                subtopic_models.SubtopicPageCommitLogEntryModel,
+                ['subtopicpage-%s-%s'
+                 % (item.id, version) for version in python_utils.RANGE(
+                     1, item.version + 1)]),
+            base_model_validators.ExternalModelFetcherDetails(
+                'snapshot_metadata_ids',
+                subtopic_models.SubtopicPageSnapshotMetadataModel,
+                snapshot_model_ids),
+            base_model_validators.ExternalModelFetcherDetails(
+                'snapshot_content_ids',
+                subtopic_models.SubtopicPageSnapshotContentModel,
+                snapshot_model_ids),
+            base_model_validators.ExternalModelFetcherDetails(
+                'topic_ids', topic_models.TopicModel, [item.topic_id])]
+
+    @classmethod
+    def _get_custom_validation_functions(cls):
+        return []
+
+
+class SubtopicPageSnapshotMetadataModelValidator(
+        base_model_validators.BaseSnapshotMetadataModelValidator):
+    """Class for validating SubtopicPageSnapshotMetadataModel."""
+
+    EXTERNAL_MODEL_NAME = 'subtopic page'
+
+    @classmethod
+    def _get_model_id_regex(cls, unused_item):
+        return '^[A-Za-z0-9]{1,%s}-\\d*-\\d*$' % base_models.ID_LENGTH
+
+    @classmethod
+    def _get_change_domain_class(cls, unused_item):
+        return subtopic_page_domain.SubtopicPageChange
+
+    @classmethod
+    def _get_external_id_relationships(cls, item):
+        return [
+            base_model_validators.ExternalModelFetcherDetails(
+                'subtopic_page_ids',
+                subtopic_models.SubtopicPageModel,
+                [item.id[:item.id.rfind(base_models.VERSION_DELIMITER)]]),
+            base_model_validators.ExternalModelFetcherDetails(
+                'committer_ids', user_models.UserSettingsModel,
+                [item.committer_id])]
+
+
+class SubtopicPageSnapshotContentModelValidator(
+        base_model_validators.BaseSnapshotContentModelValidator):
+    """Class for validating SubtopicPageSnapshotContentModel."""
+
+    EXTERNAL_MODEL_NAME = 'subtopic page'
+
+    @classmethod
+    def _get_model_id_regex(cls, unused_item):
+        return '^[A-Za-z0-9]{1,%s}-\\d*-\\d*$' % base_models.ID_LENGTH
+
+    @classmethod
+    def _get_external_id_relationships(cls, item):
+        return [
+            base_model_validators.ExternalModelFetcherDetails(
+                'subtopic_page_ids',
+                subtopic_models.SubtopicPageModel,
+                [item.id[:item.id.rfind(base_models.VERSION_DELIMITER)]])]
+
+
+class SubtopicPageCommitLogEntryModelValidator(
+        base_model_validators.BaseCommitLogEntryModelValidator):
+    """Class for validating SubtopicPageCommitLogEntryModel."""
+
+    EXTERNAL_MODEL_NAME = 'subtopic page'
+
+    @classmethod
+    def _get_model_id_regex(cls, item):
+        # Valid id: [subtopicpage]-[subtopic_id]-[subtopic_version].
+        regex_string = '^(subtopicpage)-%s-\\d*$' % (
+            item.subtopic_page_id)
+
+        return regex_string
+
+    @classmethod
+    def _get_change_domain_class(cls, item):
+        if item.id.startswith('subtopicpage'):
+            return subtopic_page_domain.SubtopicPageChange
+        else:
+            cls._add_error(
+                'model %s' % base_model_validators.ERROR_CATEGORY_ID_CHECK,
+                'Entity id %s: Entity id does not match regex pattern' % (
+                    item.id))
+            return None
+
+    @classmethod
+    def _get_external_id_relationships(cls, item):
+        return [
+            base_model_validators.ExternalModelFetcherDetails(
+                'subtopic_page_ids',
+                subtopic_models.SubtopicPageModel,
+                [item.subtopic_page_id])]
 
 
 class UserSettingsModelValidator(base_model_validators.BaseUserModelValidator):
@@ -4689,121 +4765,3 @@ class PlatformParameterSnapshotContentModelValidator(
                 [item.id[:item.id.find('-')]]
             )
         ]
-
-
-class SubtopicPageModelValidator(base_model_validators.BaseModelValidator):
-    """Class for validating SubtopicPageModel."""
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        return '^%s-\\d*$' % (item.topic_id)
-
-    @classmethod
-    def _get_model_domain_object_instance(cls, item):
-        return subtopic_page_services.get_subtopic_page_from_model(item)
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        snapshot_model_ids = [
-            '%s-%d' % (item.id, version) for version in python_utils.RANGE(
-                1, item.version + 1)]
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_commit_log_entry_ids',
-                subtopic_models.SubtopicPageCommitLogEntryModel,
-                ['subtopicpage-%s-%s'
-                 % (item.id, version) for version in python_utils.RANGE(
-                     1, item.version + 1)]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'snapshot_metadata_ids',
-                subtopic_models.SubtopicPageSnapshotMetadataModel,
-                snapshot_model_ids),
-            base_model_validators.ExternalModelFetcherDetails(
-                'snapshot_content_ids',
-                subtopic_models.SubtopicPageSnapshotContentModel,
-                snapshot_model_ids),
-            base_model_validators.ExternalModelFetcherDetails(
-                'topic_ids', topic_models.TopicModel, [item.topic_id])]
-
-    @classmethod
-    def _get_custom_validation_functions(cls):
-        return []
-
-
-class SubtopicPageSnapshotMetadataModelValidator(
-        base_model_validators.BaseSnapshotMetadataModelValidator):
-    """Class for validating SubtopicPageSnapshotMetadataModel."""
-
-    EXTERNAL_MODEL_NAME = 'subtopic page'
-
-    @classmethod
-    def _get_model_id_regex(cls, unused_item):
-        return '^[A-Za-z0-9]{1,%s}-\\d*-\\d*$' % base_models.ID_LENGTH
-
-    @classmethod
-    def _get_change_domain_class(cls, unused_item):
-        return subtopic_page_domain.SubtopicPageChange
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_ids',
-                subtopic_models.SubtopicPageModel,
-                [item.id[:item.id.rfind(base_models.VERSION_DELIMITER)]]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'committer_ids', user_models.UserSettingsModel,
-                [item.committer_id])]
-
-
-class SubtopicPageSnapshotContentModelValidator(
-        base_model_validators.BaseSnapshotContentModelValidator):
-    """Class for validating SubtopicPageSnapshotContentModel."""
-
-    EXTERNAL_MODEL_NAME = 'subtopic page'
-
-    @classmethod
-    def _get_model_id_regex(cls, unused_item):
-        return '^[A-Za-z0-9]{1,%s}-\\d*-\\d*$' % base_models.ID_LENGTH
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_ids',
-                subtopic_models.SubtopicPageModel,
-                [item.id[:item.id.rfind(base_models.VERSION_DELIMITER)]])]
-
-
-class SubtopicPageCommitLogEntryModelValidator(
-        base_model_validators.BaseCommitLogEntryModelValidator):
-    """Class for validating SubtopicPageCommitLogEntryModel."""
-
-    EXTERNAL_MODEL_NAME = 'subtopic page'
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        # Valid id: [subtopicpage]-[subtopic_id]-[subtopic_version].
-        regex_string = '^(subtopicpage)-%s-\\d*$' % (
-            item.subtopic_page_id)
-
-        return regex_string
-
-    @classmethod
-    def _get_change_domain_class(cls, item):
-        if item.id.startswith('subtopicpage'):
-            return subtopic_page_domain.SubtopicPageChange
-        else:
-            cls._add_error(
-                'model %s' % base_model_validators.ERROR_CATEGORY_ID_CHECK,
-                'Entity id %s: Entity id does not match regex pattern' % (
-                    item.id))
-            return None
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'subtopic_page_ids',
-                subtopic_models.SubtopicPageModel,
-                [item.subtopic_page_id])]
